IJPL-185469 Rename block terminal feedback stuff
To adapt and reuse this code for the reworked terminal,
let's start by renaming the existing classes and changing
the feedback ID.

Create a separate local storage for the state
we need for the feedback form.
For now, drop the most used shell
and the executed command number because
it's not clear at this point where to get these from.
We can always re-add them back.

GitOrigin-RevId: 44fd74472c60909ebc6ed1be45f8c60356dfb10a