[terminal] IJPL-185018 Enable the Reworked Terminal by default for new users
We first remember that the user is new by detecting it on startup using `AppLifecycleListener` that is implemented by `TerminalNewUserTracker`.
Then, when a user opens the terminal for the first time (it can be not on the first IDE startup), we will set the Reworked Terminal engine instead of the Classic one during `TerminalOptionsProvider` init.

Perform the terminal engine change lazily, because it is not a good idea to call terminal-related classes on startup and cause additional classloading. Also, we perform forceful saving of the settings after initialization in `TerminalOptionsProvider`, that may negatively affect the startup.

(cherry picked from commit 48f2907eb00dc608aee1825e6191ffafb2ad6e44)

IJ-CR-161382

GitOrigin-RevId: 3b06d9937fcc84f950c14de217a83155d34af435