IJPL-185469 Show reworked terminal feedback notification on engine switch
When switching away from the reworked engine, show that notification.
To make it work, the var property had to be converted to
val + switchTerminalEngine because it's needed to pass the project
to show the notification.

The corresponding test setter had to be moved into the platform.

(cherry picked from commit 5d90d02879cbe66f78704971a8d48f601ba466ce)

IJ-CR-161379

GitOrigin-RevId: 99e7f01f8f310bff2d9d12f3bbb29e5287c41792