[JEWEL-809] Fix typo in macOS decorated window property
In decorated window, on macOS, we are setting the
`apple.awt.newFullScreeControls` system property. It used to be spelled
this way in IJP.

However, it was fixed in IJP in early 2024 to
`apple.awt.newFullScreenControls`. This updates the system property,
fixing the window controls in fullscreen.

This also updates ui.api, as it seems to be out of date.