IJPL-185469 Count terminal enter key presses on the backend
This is needed for the feedback notification to work properly
when the terminal is disabled through the settings
because the settings code is invoked on the backend through Lux.

The isRemoteDevHost() check is needed to avoid counting
enters twice in the monolith.

(cherry picked from commit b45e46ba0d2ec0cd60c0ac3ce96ae7a7e619a853)

IJ-CR-161379

GitOrigin-RevId: 2d84efa6bd09ba289ba9c621c9e8babf5b74171c