[rdct] fix error from test "OpenAPI modules must not depend on implementation-detail libraries"
junit.framework.AssertionFailedError: OpenAPI module 'intellij.platform.kernel' depends on 'fastutil-min' library. This is bad because this library is an implementation detail which must not be exposed in Open API. Please move this dependency down to the corresponding 'impl' module, or use similar classes from the JDK instead.

fastutil-min is required there for `ReadTrackingIndex` class, and it's not used anywhere.
So, deleting the class to comply with the test requirements

GitOrigin-RevId: 26a95b5c3b95847f336f9ca82d237e8876c14b4c