PY-79143: Refactor Logical Structure mapping to AI
Use BFS instead of DFS for mapping. It improves YAML structure (it becomes more flat) and allows limiting the number of nodes.
If the node is already described - it will leave only the name of the node without duplicating the properties.
Move all code about mapping to the llm plugin.
Dump Logical Structure action has been moved as well and marked as internal.

GitOrigin-RevId: 04713bc2a67dd8dbb17536614032f5bdc5e5c1d5