[recent files] Introduce backend model that accumulates changes received from all the registered listeners and combines them into an end-user-visible files list
This is required to be able to filter out irrelevant changes, e.g. when a file not present among recently opened files is modified. Thus the model now consists of the gateway that processes all changes and the state that stores the resulting list. On the frontend side only the resulting state is present as well, and it is accumulated just the same way as on the backend side.

GitOrigin-RevId: 5657dd604b7fec128481c76d452e4c31d3af22c2