IJPL-185469 Create TerminalFeedbackAction
An action to explicitly send feedback.

The condition to show the action is based only on
checkIdeIsSuitable() because it's assumed that
the user may want to send updated feedback
even after it was already sent.

Convert ReworkedTerminalSurveyConfig
to object to avoid creating new instances of it
on every action update.

The action had to be registered in the frontend
module because it's where the tool window action
group is. It's impossible to use add-to-group
instead because when the action is registered,
the group isn't created yet (the frontend
module depends on the shared module and therefore
is loaded later).

GitOrigin-RevId: 5252f5d072e547b3bbb075829e6c37d03322289e