[platform] don't fail to start if a resource bundle referenced from some plugin cannot be found (IJPL-185981)
After 005f0b5daac1c665, precomputed 'bundle' isn't used: if it's not null, 'bundleName' is not null as well, so 'bundleSupplier' won't use 'bundle' and find it by its name instead. And it won't catch 'MissingResourceException', so if the bundle cannot be found, the IDE fails to start. Now the unused code is removed, and 'MissingResourceException' is caught and reported as an error as before.


(cherry picked from commit 5775e10401c940a5b6e82e20daaf9dfc5ce9fd89)

IJ-CR-161681

GitOrigin-RevId: aa58b078c99b43676941c68ad054b5fe522786ce