[terminal] IJPL-185658 Fix restoring terminal tabs on startup when the classic terminal is enabled
The condition based on the client app session was always true, so we always skipped the tabs restoring. Even in the monolith case.
Fixed it for the monolith case now.
But in the RemDev case, tabs are still not restored correctly after full restart. Only after the client reconnect.
Fixing it for the RemDev case is more complicated and should be out of the scope of this issue.

GitOrigin-RevId: 7ebee9d76b3b763cd0f6a3302fa3ffd42e1fd2fe