KTIJ-33972 [kotlin] Properly handle imports of `component` functions in `for` loop destructuring
For `ComponentFunctionMissing`, simply rely on the attached `destructingType`.

For `ComponentFunctionAmbiguity` we have to rely on Analysis API to get the right type,
see KT-77203.

In that case, detect when we are in the `for` loop destructuring via PSI
and get the type based on the entire `loopParameter`.

^KTIJ-33972 Fixed


(cherry picked from commit b1fe1423f666be4e83b980cbc04f8341035a7b72)

IJ-CR-161638

GitOrigin-RevId: f56ee088c610d6d0f1fe0abf293c7e62b59d1553