[rdct] IJPL-183208: Fix faulty condition about tool windows
The condition for being inside the tool window was incorrect. The part added in IJPL-169760 meant to check that in RD scenario the focus is not on the tool window's parent. However, the written condition was incorrect. That led to the situation where the structure view had tried to refresh itself on the popup opening. That led to the incorrect results.

GitOrigin-RevId: c8888ca7de3f5bd5bc79a14fdcf92630688e805f