IJPL-186161 Fix terminal pinch zoom
It's a bit tricky because the magnificator API requires
us to return the zoom point. So we have to encapsulate
it into the strategy. Another tricky part is that the point
we already have is relative to the viewport,
but the one needed here is relative to the whole editor.
So we need to convert.

GitOrigin-RevId: 4a98544f73ab5722448d837c8f7ef63f2712e5d4