[platform] mention ID of the problematic plugin if resource bundle is missing (IJPL-185981)
After 21cf4386 DynamicBundle.getResourceBundle doesn't throw MissingResourceException anymore if the bundle is missing, so instead of catching it we need to report proper PluginException in DynamicBundle.

PluginException is not accessible from AbstractBundle, so the old behavior is kept there. It shouldn't be a problem since code which uses AbstractBundle.getResourceBundle will probably have classes from the corresponding plugin in the stacktrace anyway.

GitOrigin-RevId: e324457ee2a391ac4c22c57aa06a48e5d458e037