IJPL-184019 Introduce TerminalPanelMarker to track terminal focus
When the terminal is placed into an editor tab,
it becomes indistinguishable from an editor.
There is no common terminal-specific component
in this case that would work for all three engines.
Introduce a marker interface so we can detect
focus on such a terminal without digging too deeply
into the internals of Swing components.

GitOrigin-RevId: c7d44cdf5876ad809680d8dadd8e59f829e1e177