[intellij project configuration] remove exceptions for Android modules from the 'Suspicious Package-Private Access' inspection (IJPL-149126)
There are several reports in Android modules, but it's better to fix them in the Google repo to avoid problems with merges. We can remove exceptions for Android modules from the inspection settings without waiting until it's done because android sources are excluded from 'Zero Tolerance' inspections for now. Removing all exceptions makes it simpler to write a test which will check that no new exceptions are added.

GitOrigin-RevId: a272221c6a6b03afd3ea09a5d2050c76fbcd2450