[java] remove UI for configuring groups of modules loaded together in 'Suspicious package-private access' inspection (IDEA-200047, IJPL-149126)
We got rid of such exceptions in intellij monorepo (see IJPL-149126), and it's better to remove the ability to add them back. Given that the inspection is currently visible in internal mode only, it shouldn't affect users. Moreover, since the usability of UI was questionable, after removing it we can make the inspection non-internal.

The option itself and code which reads its configuration are kept to avoid red code in older branches of the monorepo where the problems in code aren't fixed.

GitOrigin-RevId: 2bf42ade9ec35de6571a1d42889569790d0655b5