[Gradle|BuildView] fix: add multiple build failure line to task output end markers
'FAILURE: Build completed with 2 failures.' happens regularly now, e.g.
when several tasks failed, but this line is not recognized currently.
As a result all this build failure message is appended to last task's
and is not parsed by the root parser resulting in completely broken
presentation on the Build Output UI.

'[Incubating] Problems report...' also does not belong to any task
so add it as well.

closes https://github.com/JetBrains/intellij-community/pull/3041

### Issues
  * IDEA-364738 IDEA does not expect MultipleBuildFailures from Gradle
  * IDEA-368882 Gradle output parser doesn't handle "FAILURE: Build completed with 2 failures" output
  * IDEA-368488 Gradle compiler error build event contains inconsistent parentId task name

GitOrigin-RevId: c6d56e40f15906e7fb684b73b65af2e3be7e7e22