IDEA-364738 Add multiple build failure line to task output end markers
'FAILURE: Build completed with 2 failures.' happens regularly now, e.g.
when several tasks failed, but this line is not recognized currently.
As a result all this build failure message is appended to last task's
and is not parsed by the root parser resulting in completely broken
presentation on the Build Output UI.

'[Incubating] Problems report...' also does not belong to any task
so add it as well.