IJPL-184019 Initialize TerminalFocusFusService with the toolwindow
To make it work for all terminal engines, initialize it
when the terminal tool window is initialized.

Because it's initialized in the shared module,
convert the light service to a regular one
and use serviceOrNull() to avoid trying to initialize it
on the backend.

(cherry picked from commit 3f193e0684b4c9f920093b56418e90cf5854ab81)

IJ-CR-161650

GitOrigin-RevId: 9d34514dd136e224ed833c4682108e3859dea317