IJPL-186424 Don't show zoom indicators for invisible editors
In the terminal, there are two editors that are zoomed in sync:
the regular buffer and the alternate buffer ones. Only one of them
is visible. Trying to show a popup for an invisible editor doesn't
work and effectively makes it invisible.

Fix by only showing zoom indicator for visible editors.

GitOrigin-RevId: 170e34d252eaf094a78958eda75a1084340c43cd