PY-79816 Ignore `generics_self_attributes.py` in conformance test suite
The test was passing for the wrong reasons. On the line

xs = OrdinalLinkedList(value=1, next=LinkedList[int](value=2))  # E

`value=2` was highlighted as an unexpected argument (which is a false-positive). PY-79816 fixes this false-positive.

However, the false-negative for highlighting `next` as an error is still there. So the test doesn't pass now

GitOrigin-RevId: 518a045e9521ba973e3c96f0a31b0ec28e0fc9f6