[AutoSync] fix: isolate project sync tracking per each external system project
The initialization process of different build tools may differ. It causes the different project activation time. Which causes the second sync cancellation if the first sync is in progress. Therefore, we need to track project syncs separately for each external system project.

The several syncs aren't allowed for one project. Therefore, project tracker cancel and reschedule secondary sync if there is a sync already in progress.

### Issues
  * IDEA-372032 Gradle auto-import not working in Maven presence

GitOrigin-RevId: 23eca0def658bc63b30f1cd7eca18e8a345059fa