IJPL-185482 Ignore outdated focus gained events in the tool window manager
Sometimes, when a modal dialog is closed, the focus
quickly changes inside the dialog. For example,
it happens with the Create Branch dialog
when the editor inside it is disposed, the focus is
transferred to some other component then.

Because focus events arrive asynchronously,
they can get to the tool window manager later,
when the modal dialog even no longer exists.
It makes no sense to hide Dock Unpinned
tool windows when such a focus change happens.

GitOrigin-RevId: 43168889ee77e8c9854cd56a7bf8df60d76898ad