[remote dev] migrate frontend plugins from a previous version when starting a new major version (IJPL-183871)
When the frontend process starts, 'JetBrainsClientConfigImportSettings.processPluginsToMigrate' will now collect frontend plugins installed in the previous IDE version and migrate them to the new one. It may also happen that a regular IDE of the new version starts first. In that case, the frontend plugins from the previous IDE version will be stored in a temporary 'frontend-to-migrate' subdirectory. And when the frontend variant of the new version starts, 'FrontendProcessPathCustomizer' will initiate plugin migration, and it'll migrate them from there. After the frontend process starts, 'CodeWithMeCleanupManager' will remove the temporary directory.

GitOrigin-RevId: a6670c5dda185eb102c7d992b27d386378ea274c