[recent files] IJPL-186498 Do not add any new file to the recent files model. Prefer listening only to editor typing, not VFS
Recently changed files model was previously subsribed to VFS changes and did consume any file added to the VFS snapshot. This is incorrect behaviour, since in our understanding the recently changed files list must only contain files edited manually. Example: if a code generation tool produces N files in the project, they should not appear in the model. But if a user types in one of them, this exact file should.

The implementation now relies on frontend typing listener with a reasonable debounce (1 sec) that must cover only expected cases

GitOrigin-RevId: b9348336aead5171682cdb70fa52cdd08b8f8214