IJPL-187025 Fix ProjectUtil.getActiveProject returning a project from a closed window
There are several possible fixes. We could check if the project
itself is disposed, for example. But because the idea of
the "active" project is closely related to the "active" window,
the simplest way is to check whether the window itself is
disposed, and we already have an API for that.

GitOrigin-RevId: 89424d17824276b2532123cbc63db5ed094f3d3a