[settingsSync] Invalidate communicator on B&S disable
Also add the missing listener registration.

This is a temporary patch to address an immediate issue. It is a
subset of the changes planned in a01d0e452900e4cfb0e663ec9d67b1ac556c65f3 and will become obsolete once that is merged.

Note that there's code conflicts between the master and release branches that prevent a clean cherry-pick from master.

Change-Id: Ieedf5e6fc9cac69f222c31c1a23c5a5a109614ec