[remote dev] cleanup: remove reference to obsolete 'com.intellij.gateway' plugin alias from plugin configurations
There is no such plugin alias since 176bb19c5de778ad4. The Gateway as a product now declares `com.jetbrains.gateway` alias, and the same ID is used for Gateway as a plugin bundled with many IDEs, so these plugins shouldn't be incompatible with it.

This is also needed to write a test which verifies validity of 'incompatible-with' tags.

GitOrigin-RevId: 8e87f29c63cf725831bf32025843bf609b0adc38