IJPL-185365 Fix tool window view mode--auto hide matching
Some leftover of the old implementation of tool window modes
left auto-hide untouched when the tool window is switched
to any mode but Dock Pinned/Unpinned. This means that
when switching from Dock Unpinned (auto-hide on) to, e.g.,
Float, the tool window would remain in the auto-hide mode.

Fix by explicitly setting the auto-hide mode for all modes.
For Dock Unpinned and Undock it's on, for the other modes
it's off.

GitOrigin-RevId: c94980bc85428b5f96c6a7a4ebfee65aa97aecd0