[eel] use 'loading=embedded' for 'intellij.platform.eel.impl' module (IJPL-187171)
IjentNioFileSystemProvider from 'intellij.platform.ijent.community.impl' module which is loaded by the core classloader refers to 'EelFsResultImpl' class from 'intellij.platform.eel.impl' module, so the latter must be loaded by the core classloader as well to avoid NoClassDefFoundError.

It would be better to register 'intellij.platform.ijent.community.impl' as a content module and add dependency on 'intellij.platform.eel.impl' instead, but this would require adding dependencies on 'intellij.platform.ijent.community.impl' as well, so it's better to do this after we have a test which verifies that (see IJPL-171921).

GitOrigin-RevId: a9a35dd921e145846ffccc175cf04ff012b3b190