[gradle] cleanup: remove unused dependencies from intellij.gradle modules
They aren't used in code and won't work at runtime anyway because there is no corresponding dependency in the plugin descriptor. A direct dependency is added to the test module instead to allow running tests in its classpath.

This change is also necessary to avoid failures in the test which verifies dependencies (IJPL-171921).

GitOrigin-RevId: 5be4ac920fb166498b4de0bccaf21abc07c28118