IJPL-187312 Add a link from the terminal settings to the colors
The names necessary for selectOrEditColor to work
were exposed through an internal API.
Instead of the parent node, the first color inside
had to be used for the reasons explained in the comment.

Note: in remdev the configurable is opened,
but nothing is selected in the tree.
No wonder, given how complicated the logic there is.

GitOrigin-RevId: 965db3a976ede653a3069bdd96e251fe4be92e77