[git] cleanup: remove unused dependency from intellij.vcs.git to intellij.vcs.git.frontend
It isn't used in code and won't work at runtime anyway the former is part of the main module of Git plugin, and the latter is an optional content module, so the dependency goes in the opposite way.

This change is also necessary to avoid failures in the test which verifies dependencies (IJPL-171921).

GitOrigin-RevId: 4e52283883cfe0604970feddd1053c1d8b9966f5