[JEWEL-817] Fix iconClass defaults for *IconActionButton
We were passing the IconKey's class itself as key instead of the class
it contains, causing issues in IDE release builds due to classloader
isolation. Now we pass the correct class in by default.

We also deprecate a few now-unused APIs related to icon loading, and
address a minor issue in the ComboBoxes showcase page.