IJPL-187512 [terminal] Implement font fallback
Because some fonts are missing some characters
used in popular shell themes, we need to have
a fallback font.

Since a fallback font can only be provided to the editor
through the FontPreferences API, we can't use
TerminalFontSettings directly anymore.
Still, we need it for TerminalOptionsConfigurable
because using FontPreferences there would be messy.

TerminalFontSettingsService already inherits
getFontPreferences from AppFontOptions,
but it only has a fallback family if it's not null
when the state is loaded (see copyState).
But we need it to be always available
except when it's not specified at all AND
the main font is also the default one.

Therefore, we override getFontPreferences
to ensure that a fallback is always provided
if necessary.

(cherry picked from commit 905fe55100ea8358fd9bea7679ccff84629da996)

IJ-CR-162755

GitOrigin-RevId: d923979fc21403f8c2e26fa23e023a2f924f5840