[terminal] IJPL-163552 Refactor: allow collecting terminal session output flow by several collectors
It is required for the new backend collector that tracks the current working directory.
So, now we explicitly collect the output events from the original flow, update the backend models and forward them to the frontend.
Surprisingly, it was also working like this before. Because RPC logic collects the flow on its side. And, for some reason, this behavior differs in split mode and the real RemDev (like CodeCanvas).
I decided to skip the investigation of it, and just leave the new behavior, to make it explicit.

GitOrigin-RevId: 779eb81d21438ef6b3e27c31b98aac0f10b4eed9