IJPL-187617 [terminal] Don't show the zoom indicator if the font size is reset
The font size can be reset for many reasons, and it's not easy
to track them all from a listener. But if it's reset, we almost
never need a zoom indicator. On the other hand, if it's explicitly
set to any value, the zoom indicator is almost always useful.
So it seems like a good approximation to determine
whether we need it or not.

Pass a boolean flag to listeners so they can decide whether
a zoom indicator is needed. We don't have this zoom
in execution-impl, so there it doesn't matter what we pass.

(cherry picked from commit eebc12f64d0aed8f5d8a39e1606193c5b704ad5b)

IJ-CR-162754

GitOrigin-RevId: 72fd686a10799ee72b1dc42daa0471d143e62c9d