[evaluation plugin] simplification: get rid of auxiliary intellij.evaluationPlugin.ultimate module
Since IJPL-176806 is implemented, it's possible to register content modules defined in the ultimate part of the project directly in plugin.xml located in intellij-community. Such modules will be skipped when building the plugin from intellij-community source code. Removal of the auxiliary intellij.evaluationPlugin.ultimate module simplifies the project configuration and build scripts.

This is also necessary to avoid failures in the test which verifies plugin dependencies (IJPL-171921).

GitOrigin-RevId: 1daa981b60e9bd0fda8f2477cfdbb1fecba9fa15