IJPL-187622 Update the active window list when a window is closed
Normally if a window is closed, another one is activated
soon, and therefore the closed window is removed
when updateActivatedWindowSet is called from
addActiveWindow.

However, in tests it's possible that a window is closed
without any focus / activation events (an inactive window is closed).
This creates false positive project leaks, as the window
remains in the list and the project remains reachable.

Because window activation is already handled in IdeEventQueue,
we can easily handle window closing there as well,
updating the list when any window is closed.

GitOrigin-RevId: a28d790b76ad64fefdf8c4911d24237c307fb528