IJPL-187695 Make launchOnceOnShow restart the job if it's canceled before starting
The implementation is as simple as a while() loop
with one caveat that we now need to wait for the UI
coroutine to finish after we cancel it.
Otherwise we wouldn't know if it managed to start or not.

Add a test for this behavior. For some reason,
yield() is not enough to make sure the job is executed,
so we use join() instead, which works fine because
the test job is just a single assignment.

GitOrigin-RevId: b649c5a813a763420ef9b3a32c3d90fd0701c365