[plugin model tests] add a test which verifies the dependencies in JPS module are satisfied at runtime (IJPL-171921)
IdeaUltimatePluginModuleDependenciesTest uses information from *content.yaml files to create plugin descriptors similar to actual ones without building the installation. Then it reuses the production code from the plugin model to set up dependencies and create classloaders for these descriptors. And after that it checks that if module 'a' has 'Compile' dependency on module 'b' in a.iml file, then the parent classloaders for 'a' must include the classloader for 'b'. If it doesn't, it reports an error and suggests possible ways of fixing it.

Problems which are currently reported are specified in KnownProblemsInModuleDependencies.kt to temporarily ignore them until they are fixed.

GitOrigin-RevId: c7a8dc15d751c85bd7299b28939483236f75526e