[JEWEL-812] Implement JewelBridgeClipboard
In the bridge only, this change adds a new Clipboard implementation that
delegates to the IJP's ClipboardManager instead of the AWT Clipboard
(which is what CMP does).

This allows us to directly integrate in the clipboard history feature in
IJP, and take advantage of a number of improvements around pasting from
the clipboard. For more details see IJP's ClipboardSynchronizer.