IJPL-187714 Fix project icon update
The cause of the problem, as explained in IJPL-188019,
is that icons with equal deferrer keys are considered equal.
To ensure that icons created after refreshing the icon
are not equal to those created before,
we introduce a counter that's incremented every time
an icon is refreshed. This counter then becomes
a part of the key along with all information
needed to load the icon.

To ensure that the code loading icons doesn't
have access to any information that's not a part of the key,
extract this code into the keys themselves.

GitOrigin-RevId: bf0fcc1cafc3d822dc2b38aa7be057bded5891ee