[terminal] IJPL-187899 Ensure that rename tab requests are executed in the same order they are called on the frontend
RPC does not guarantee that methods called in some order on the frontend will be executed in the same order on the backend. So, for methods that are not idempotent, we need to invent some ordering mechanisms.
Ideally, we need to use the `SendChannel` in RPC for this, like we do in `TerminalSessionApi`, but here it looks like an overkill for now.
So, use the channel to buffer the requests and send them sequentially one after another.

GitOrigin-RevId: 9b6a448a70185ec0b9298c850b6c9a2d9cc5f55a