IJ-CR-160632 Fix yet another SAM listener bug
Classics. The instances passed to addEditorHyperlinkListener
and removeEditorHyperlinkListener are different because
the Kotlin function is converted to the Java interface every time.
Not critical in this case, but better be fixed, if only
to not serve as a dangerous example for others.

(cherry picked from commit 11295d27ff2591610425051840bdb73509db960f)

GitOrigin-RevId: 483cf9950617e55756bb08898fa7af6a2bd1ac8a