ML in SE: Handle potentially missing typos module gracefully (IJPL-183114)
It may happen that the typos submodule is missing, this leads to IllegalArgumentException
getting thrown when getting advanced settings - since the key is defined in the submodule.
In order not to break the whole Search Everywhere (like in the IJPL-183114 issue)
we are going to catch the exception, log the error, and return no experiment.

GitOrigin-RevId: b09a2999d04631c009ec9cb558def4722f77ef0a