IJPL-188109 [terminal] A custom color scheme delegate for the terminal
This is the-least-of-all-evils hopefully temporary solution
for all the problems that are caused by the interaction
between the terminal settings, the editor, its color scheme
delegate (EditorImpl.MyColorSchemeDelegate)
and the global color scheme.

Introduce TerminalColorScheme, a safe wrapper
around the global color scheme.
It allows us to delegate what we don't override
and to override what we need.
The global scheme is updated when it's changed.

GitOrigin-RevId: dabaa846cb905b58d191ee8753c91d7a646ecc3d