[Gradle|Exec] fix: execute only one project build task by Gradle in at the same moment
Only one project build task is allowed to be executed by Gradle at the same moment.
It is necessary to reduce the number of working Gradle daemons in parallel.

The GradleProjectTaskRunner API allows running several project build tasks in a batch.
However, they should be provided as one run transaction.

### Issues
  * IDEA-359288 Cannot run Gradle compound run configurations because of fighting due to multiple daemons

GitOrigin-RevId: b64bdd1c42f99bdaff93fa81155075826489979b