[git] IJPL-173922 Move actions from branches widget settings to shared module
`GitUiSettingsApi` is used for updating `GitVcsSettings` values as updates are not automatically propagated from frontend to avoid problems with listeners invocations in setters.

It's also possible to get rid of some methods in `GitUiSettingsApi` by extracting the frontend-specific part of `GitVcsSettings`. However, there are still methods as `setGroupingByPrefix` or `setShowTags` which should invoke listeners on the backend

GitOrigin-RevId: d85916ed1b0e32d8e7910b3ef3a42119c669f185