[terminal] IJPL-188222 Use TerminalSessionsManager in tests
So, we will test the same backend terminal session that is used in production. With both JediTerm and `StateAwareTerminalSession` parts.
Previously, we tested only JediTerm part. And missed a bug with missed `TerminalSessionTerminatedEvent` event because if that.
Use `StateAwareTerminalSession` only in the tests that do not require low-level checks.

(cherry picked from commit 3f8dfd607682e06153b4731b0fbb310addf9b5ed)

IJ-CR-163462

GitOrigin-RevId: e8ff2e88d95d50f30f119fce9cf01716b886670a