[terminal] IJPL-188222 Ensure that termination event is emitted when session is terminated using interaction with shell
Like `exit` command or Ctrl+D shortcut.
It was not delivered because after emitting the event, we immediately cancelled the coroutine scope in `StateAwareTerminalSession` that collected the events and forwarded them to the frontend.
Now JediTerm session is running in a separate scope, so its cancellation is not affecting the `StateAwareTerminalSession` scope.
While in `StateAwareTerminalSession` we cancel the main session scope if we receive the `TerminalSessionTerminatedEvent`.

(cherry picked from commit 6f97981ef84f16c1a181462bc054e15a0e4134fc)

IJ-CR-163462

GitOrigin-RevId: 9ff03ff440d1bab059f1cf90f181d3e757fcc65f