syntax: add support for custom whitespaces in xml
XML has a special whitespace kind `XmlTokenType.XML_REAL_WHITE_SPACE`.
On AST construction, a leaf with this token type gets replaced with a plain PsiWhiteSpaceImpl
(which has the plain WHITE_SPACE token type), see `XmlASTFactory.createLeaf`,
so XML_REAL_WHITE_SPACE exists during parsing only.

But when we want to reparse the file, we end up in a situation when a new not-yet-built tree is
compared with the existing tree. And here, we need to deal with inconsistency in token types between XML_REAL_WHITE_SPACE in the new tree and WHITE_SPACE in the old tree.

To overcome this stupid situation, this ugly hack is introduced.

GitOrigin-RevId: 2f214e6aaf98923df88bf76aceff14ce6c4f748d