[jps model] add JpsProject.findModuleByName method
After IDEA-368571 is fixed, it's possible to quickly find a module by its name, but it wasn't exposed in API, and we still have many places in code which iterate over all modules to find a module by name.

Also, JpsModuleReferenceBridge is changed to use this new method to resolve module references, so we won't face problems like IDEA-368571 again after switching to the new implementation based on the workspace model.

GitOrigin-RevId: 04b2d58bd21258139f9e5e20a9f96c76b17b1d6f