IJPL-177442 Moved backend plugin operations to a separate class.
DefaultUiPluginManagerController is the default implementation for plugin operations. It will be used in case of the default registry key value and for the 'frontend' implementation in the future. Backend calls are also delegated to this class. Most of the state has been moved to the backend as well.


(cherry picked from commit 0bf8c684d82f5a535190bb1e95bfe17986def0c2)

IJ-MR-163725

GitOrigin-RevId: 74cbc74eaf272be87df4b073e077b6e22a9f76bf