[plugin model] refactoring: provide a way to specify custom product mode for plugin loader in tests (IJPL-171921)
This is needed to check plugin module dependencies for frontend variants of IDEs. The function which evaluates the current product mode is now passed to PluginSetBuilder::computeEnabledModuleMap, and can be taken from PluginInitializationContext. It would be better to take it from PluginInitializationContext in all cases, but it isn't available in the places PluginSetBuilder.createPluginSetWithEnabledModulesMap is called.

GitOrigin-RevId: 02fe3089e7ccde00888f542f5d40ac8ce2c6b9b2