[git] IJPL-173922 Propagate branches widget data context to backend
There are many actions for the git widget RDfication of which will be challenging. As an intermediate solution, backend-delegating actions and data-context propagation are used.

In split mode data context is propagated via `ActionTimestampProvider` things. Note that a git repositories model differs on the backend and frontend sides, so only the selected ref key can be reused.

In monolith `GitBranchesWidgetDataSnapshotProvider` is used.

GitOrigin-RevId: fab608f5e7b70bb16490ad0b1a163d3753216435