PY-81376 When creating a Uv interpreter, properly update module settings on EDT
Otherwise, we create a new Uv virtual env but then on actually setting module SDK
there is an exception about wrong thread access rights.
As a result, this environment needs to be set up as a regular in-project .venv
with a separate quick fix.
PyPipfileSdkConfiguration and PyPoetrySdkConfiguration already updated
settings with the correct dispatcher.

GitOrigin-RevId: 56cf6116ed8a7d41b303f013fc4db882d16ac5f9