[remote dev] use per-module classloaders when running the frontend process from source without dev build (IJPL-188953)
The frontend process uses the module-based loader, so it adds to the classpath of the core loader only modules specified in 'main-root-modules' in product-modules.xml. So to support running from sources, we currently add many modules registered as content modules to product-modules.xml as well. This lead to these modules added to the classpath to two classloaders in production. To fix the problem, now ModuleBasedProductLoadingStrategy uses a special variant of PathResolver when running from sources, which can locate modules even if they aren't registered in the product-modules.xml.

GitOrigin-RevId: c118796dd74c9d3da7c0fe791303aca05c803a3a