[terminal] IJPL-188621 Use console font preferences to initialize the terminal font settings
Found out that `AppConsoleFontOptions` for some reason do not contain the actual console font specified in the settings. It is stored in the color scheme instead, so we need to use it there.
Unfortunately, this fix will work only for the users that restore to the default IDE settings (or new users).
Since already wrong font was already stored in the terminal persistent state component. So, users still need to change the font in the terminal font settings.

GitOrigin-RevId: c0dfbe75f6c6d49a10dc9578310058c7a9f6dd6e