[jcef] IJPL-185273 Make it possible to access dev tools in the frontend browser in internal ide mode
The corresponding context menu action is disabled by default since it overrides default actions from the browser itself. It seems quite untrivial to merge these two kinds of actions together, so in the first implementation a registry flag is in charge of switching between native browser actions and custom ones.

GitOrigin-RevId: 27a6a01db2da46cb3906bca4d0e191e3ccbf8d7b