[runtime product] support 'embedded' loading in RuntimeModuleGroup (IJPL-188953)
To distinguish which modules from the main module group in product-modules.xml should be loaded by the core classloader, we need to support 'embedded' loading rule here the same way as it's supported in plugin.xml files. Also, the 'embedded' loading rule is now supported in the PluginModuleGroup for consistency.

GitOrigin-RevId: f49d118c3a889635341e9f52803124f7b7e408b4