[module-based loader] add only modules with loading=embedded to the classpath of the core classloader (IJPL-188953)
Otherwise, modules which are registered as content modules are added to classpath of two classloaders, and this may lead to problems. product-modules.xml are updated to use loading=embedded for modules which aren't registered as content modules and therefore should be added to the classpath of the main classloader of the core plugin.

'product-modules.xml' files are updated to set 'loading=embedded' where it's needed.

GitOrigin-RevId: a9cf5f5c0ddedb78fe62a47d4d91b804ee4359d6