[recent files] IJPL-188466 Do not add files to the model on selection change event
Assuming that selection changed might happen only for already existing editors, we want to make sure that if an event contains file that has not been opened, it does not get added to the model. The assumtion is based on a report from the issue that unrelated files appear in the switcher. From the logs it seems that files initially appear on the frontend side among the `frontendRecentFiles` list. The list in its turn gets assembled either from frontend editor selection history, or from a few events. Events like editor is opened/closed are not suspicious, but the selection update is. Thus, we are going to avoid adding files from the event to the model, only update those already present in it

GitOrigin-RevId: 567fa110273a17ec947696e6bb26897dc36ce9b6