[build scripts] fix paths in the 'layout' section of product-info.json for JetBrains Client distribution (IJPL-189470)
If the 'listBundledPlugins' command is started using the module-based loader, it doesn't pack modules to JARs and therefore paths to module output directories are returned in the 'layout' section of the JBC-builtinModules.json file. It's possible to convert them to real JAR paths in the distribution using information from the runtime module repository, but it's more reliable to use the launcher based on the dev build for this. This makes 'collecting compatible plugins' step a bit longer (7 seconds instead of 3 on my machine), but this shouldn't be a big problem.

GitOrigin-RevId: f74a7c49453555a1ee69edd9bc6641735110fcbc